[CI] [100460,100459]  <bt> <fix>Fix intermittent mailbox pkt_type corruption causing CP :unknown type:

+++ ap/components/bk_bluetooth/core/ipc/include/bt_ipc_core.h
@@ -60,7 +60,7 @@ typedef struct __attribute__((packed))
 typedef struct __attribute__((packed))
 {
     mb_chnl_hdr_t hdr;
-    uint8_t pkt_type;
+    uint32_t pkt_type;
 #if 0
     union
     {
